wait for micro to listen before starting browser
