+++ misc-screenshot/now.json
@@ -1,5 +1,6 @@
 {
   "type": "docker",
+  "public": true,
   "features": {
     "cloud": "v2"
   },
