+++ misc-screenshot/src/setup.js
@@ -1,8 +1,4 @@
-// const fs = require('fs');
-// const tar = require('tar');
-// const zstd = require('node-zstd').decompressStream;
 const puppeteer = require('puppeteer');
-// const config = require('./config');
 
 const isBrowserAvailable = async (browser) => {
   try {
@@ -14,38 +10,25 @@ const isBrowserAvailable = async (browser) => {
   return true;
 };
 
-// const setupLocalChrome = () => new Promise((resolve, reject) => {
-// console.log(config.chromeLibSrc);
-// fs.createReadStream(config.chromeLibSrc)
-// .on('error', reject)
-// .pipe(zstd())
-// .on('error', reject)
-// .pipe(tar.x({
-// C: config.chromeLibDst,
-// }))
-// .on('error', reject)
-// .on('end', resolve);
-// });
-
-// const setupChrome = async () => {
-// console.log('setup local chrome');
-// await setupLocalChrome();
-// console.log('setup done');
-// };
-
 exports.getBrowser = (() => {
-  let browser;
+  let browser; let
+    launching;
   return async () => {
-    if (typeof browser === 'undefined' || !await isBrowserAvailable(browser)) {
-      // await setupChrome();
+    let unlock;
+    // eslint-disable-next-line no-unused-expressions
+    (launching && launching.then && await launching);
+    // eslint-disable-next-line no-return-assign
+    launching = new Promise(r => unlock = r);
+    if (!browser || !await isBrowserAvailable(browser)) {
       browser = await puppeteer.launch({
         executablePath: '/usr/bin/chromium-browser',
-        args: ['--no-sandbox', '--headless', '--disable-gpu', '-—disable-dev-tools', '--single-process'],
+        args: ['--no-sandbox', '--headless', '--disable-gpu', '-—disable-dev-tools'],
         dumpio: true,
         devtools: false,
       });
       console.log(`launch done: ${await browser.version()}`);
     }
+    unlock();
     return browser;
   };
 })();
